test(scheduler): extract testSchedulerOption helper to dedicated file
Moves newTestScheduler and the five with* options out of integration_test.go
into scheduler/test_helpers_test.go. Per project convention (CLAUDE.md
"Build Tag Isolation"), files named *_integration*.go may later receive a
//go:build integration tag; keeping a helper used by module_test.go in such
a file is a latent compile break.

Addresses coderabbit nitpick on #362.
